Level_3/LLM_Telegram_Bot/requirements.txt
emoji==2.15.0
groq==0.32.0
loguru==0.7.3
pyTelegramBotAPI==4.29.1
python-dotenv==1.1.1
